Add Drawing Option and Option Values section
and fill out this section more
