+++ mathics/builtin/plot.py
@@ -219,155 +219,6 @@ class ColorData(Builtin):
         return palette.colors()
 
 
-# Note In defining these class, documentation is given for them
-
-
-class Axis(Builtin):
-    """
-    <dl>
-      <dt>'Axis'
-      <dd>is a possible value for the 'Filling' option.
-    </dl>
-
-    >> ListLinePlot[Table[Sin[x], {x, -5, 5, 0.5}], Filling->Axis]
-     = -Graphics-
-    """
-
-
-class Bottom(Builtin):
-    """
-    <dl>
-      <dt>'Bottom'
-      <dd>is a possible value for the 'Filling' option.
-    </dl>
-
-    >> ListLinePlot[Table[Sin[x], {x, -5, 5, 0.5}], Filling->Bottom]
-     = -Graphics-
-    """
-
-
-class Filling(Builtin):
-    """
-    <dl>
-      <dt>'Filling Top |Bottom|Axis'
-      <dd>is a an option to Plot to specify what filling to add under point, curves, and surfaces
-    </dl>
-
-    >> ListLinePlot[Table[Sin[x], {x, -5, 5, 0.5}], Filling->Axis]
-     = -Graphics-
-    """
-
-
-class Full(Builtin):
-    """
-    <dl>
-      <dt>'Full'
-      <dd>is a possible value for the 'Mesh' and 'PlotRange' options.
-    </dl>
-    """
-
-class MaxRecursion(Builtin):
-    """
-    <dl>
-      <dt>'MaxRecursion'
-      <dd>is an option for functions like NIntegrate and Plot that specifies how many recursive subdivisions can be made.
-    </dl>
-
-    >> NIntegrate[Exp[-10^8 x^2], {x, -1, 1}, MaxRecursion -> 10]
-     =  1.97519*^-207
-    """
-
-
-class ImageSize(Builtin):
-    """
-    <dl>
-      <dt>'ImageSize'
-      <dd>is an option that specifies the overall size of an image to display.
-    </dl>
-
-    Specifications for both width and height can be any of the following:
-    <dl>
-      <dt>Automatic
-      <dd>determined by location or other dimension (default)
-      <dt>Tiny, Small, Medium, Large
-      <dd>pre defined absolute sizes
-    </dl>
-    """
-
-
-class Joined(Builtin):
-    """
-    <dl>
-      <dt>'Joined $boolean$'
-      <dd>is an option for 'Plot' that gives whether to join points to make lines.
-    </dl>
-
-    >> ListPlot[Table[n ^ 2, {n, 10}], Joined->True]
-     = -Graphics-
-    """
-
-
-class Mesh(Builtin):
-    """
-    <dl>
-       <dt>'Mesh'
-      <dd>is an option for 'Plot' that specifies the mesh to be drawn. The default is 'Mesh->None'.
-     </dl>
-
-    >> Plot[Sin[Cos[x^2]],{x,-4,4},Mesh->All]
-     = -Graphics-
-
-    >> Plot[Sin[x], {x,0,4 Pi}, Mesh->Full]
-     = -Graphics-
-
-    >> DensityPlot[Sin[x y], {x, -2, 2}, {y, -2, 2}, Mesh->Full]
-     = -Graphics-
-
-    >> Plot3D[Sin[x y], {x, -2, 2}, {y, -2, 2}, Mesh->Full]
-     = -Graphics3D-
-    """
-
-    messages = {
-        "ilevels": "`1` is not a valid mesh specification.",
-    }
-
-
-class PlotPoints(Builtin):
-    """
-    <dl>
-      <dt>'PlotPoints $n$'
-      <dd>A number specifies how many initial sample points to use.
-     </dl>
-
-    >> Plot[Sin[Cos[x^2]],{x,-4,4}, PlotPoints->22]
-     = -Graphics-
-    """
-
-
-class PlotRange(Builtin):
-    """
-    <dl>
-      <dt>'PlotRange $n$ | All | Automatic'
-      <dd>is an option for 'Plot' that gives the range of coordinates to include in a plot.
-    </dl>
-
-    >> Plot[Sin[Cos[x^2]],{x,-4,4}, PlotRange -> All]
-     = -Graphics-
-    """
-
-
-class Top(Builtin):
-    """
-    <dl>
-    <dt>'Top'
-        <dd>is a possible value for the 'Filling' option.
-    </dl>
-
-    >> ListLinePlot[Table[Sin[x], {x, -5, 5, 0.5}], Filling->Axis|Top|Bottom]
-     = -Graphics-
-    """
-
-
 def extract_pyreal(value):
     if isinstance(value, Real):
         return chop(value).round_to_float()
